+++ src/scicloj/ml/tribuo.clj
@@ -3,9 +3,14 @@
   (:require
    [scicloj.metamorph.ml :as ml]
    [tech.v3.dataset :as ds]
+   [tech.v3.dataset.modelling :as ds-mod]
    [tech.v3.datatype :as dt]
+   [tech.v3.datatype.functional :as fun]
+   [fastmath.stats :as stats]
    [tech.v3.datatype.errors :as errors]
-   [tech.v3.libs.tribuo :as tribuo]))
+   [tech.v3.libs.tribuo :as tribuo])
+  (:import [org.tribuo.regression.evaluation RegressionEvaluator]
+          [org.tribuo.regression Regressor]))
 
 (defn- make-trainer [options]
   (tribuo/trainer (:tribuo-components options)
@@ -36,7 +41,6 @@
     
 
   (fn [feature-ds thawed-model {:keys [model-data target-columns] :as model}]
-    (def model model)
     (let [target-column-name (first target-columns)
           prediction
           (->
@@ -46,14 +50,86 @@
       prediction))
   {})
 
+(defn evaluate [model]
+  (let [
+        feature-ds (-> model :model-data :feature-ds)
+        target-ds (-> model :model-data :target-ds)
+
+        model-object (-> model :model-data :model)
+        evaluator (RegressionEvaluator.)
+        evaluation
+        (.evaluate evaluator model-object (tribuo/make-regression-datasource
+                                           (ds/append-columns feature-ds (ds/columns target-ds))))]
+    evaluation))
+
+(defn ->predictions [evaluation]
+
+  (flatten
+   (for [prediction (.getPredictions evaluation)]
+     (for [output (iterator-seq (.iterator (.getOutput prediction)))]
+       (for [tupple (iterator-seq (.iterator output))]
+         {:regressor (.getName tupple)
+          :prediction (.getValue tupple)})))))
+
+
+
+(defn glance-fn-regression [model]
+
+  (let [
+
+        target-ds (-> model :model-data :target-ds)
+        target-column-names (ds-mod/inference-target-column-names target-ds)
+        _ (assert (< (count  target-column-names) 2) "Can only handle single inference target")
+
+        evaluation (evaluate model)
+        predictions (->predictions evaluation)
+
+
+        regressor-values (-> target-ds :disease-progression)
+        prediction-values (map :prediction predictions)
+
+        target-column-names (ds-mod/inference-target-column-names target-ds)
+        regressor-name (name (first target-column-names))
+        regressor (Regressor. regressor-name, Double/NaN)]
+        
+
+
+    (ds/->dataset
+     {:regressor [ regressor-name]
+      :r.squared [(.r2 evaluation regressor)]
+      :mae [(.mae evaluation regressor)]
+      :rmse [(.rmse evaluation regressor)]
+      :rss (stats/rss regressor-values prediction-values)})))
+
+
+(defn augment-fn-regression [model data]
+
+  (let [prediction-values
+        (map
+         :prediction
+         (->predictions (evaluate model)))
+
+        residuos
+        (fun/-
+         (get data
+              (first (model :target-columns)))
+         prediction-values)]
+    (-> data
+        (ds/add-column (ds/new-column :.fitted prediction-values))
+        (ds/add-column (ds/new-column :.resid residuos)))))
 
 
 (ml/define-model! :scicloj.ml.tribuo/regression
   (fn [feature-ds target-ds options]
-    (tribuo/train-regression (make-trainer options) (ds/append-columns feature-ds (ds/columns target-ds))))
+    {:target-ds target-ds
+     :feature-ds feature-ds
+     :model
+     (tribuo/train-regression (make-trainer options) (ds/append-columns feature-ds (ds/columns target-ds)))})
 
   (fn [feature-ds thawed-model {:keys [model-data target-columns]}]
-    (->
-     (tribuo/predict-regression model-data feature-ds)
-     (post-process-prediction (first target-columns))))
-  {})
+    (let [model (:model model-data)]
+      (->
+       (tribuo/predict-regression model feature-ds)
+       (post-process-prediction (first target-columns)))))
+  {:glance-fn glance-fn-regression
+   :augment-fn augment-fn-regression})