+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+- added suport for glance and augment for regression 
 - call prediction column prediction
 - use latest tehc.ml.dataset
 - support :string and categorical target columns